Mejoro el código del formulario
