+++ src/screens/CreateEventScreen/CreateEventScreen.jsx
@@ -1,48 +1,18 @@
 import React from 'react'
 import './CreateEventScreenStyles.css'
-import { useForm } from '@mantine/form'
 import { Space, Title } from '@mantine/core'
 import { useNavigate } from 'react-router'
 import { notifications } from '@mantine/notifications'
 import apiProvider from '../../api/apiProvider'
-import {
-  capacityValidation,
-  locationValidation,
-  requiredField,
-} from './formValidations'
 import EventForm from '../../components/EventForm/EventForm'
 
 const CreateEventScreen = () => {
   const navigate = useNavigate()
 
-  const formState = useForm({
-    initialValues: {
-      title: '',
-      type: '',
-      dateTime: '',
-      description: '',
-      capacity: 0,
-      location: {
-        address: '',
-        latitude: '',
-        longitude: '',
-      },
-      status: 'draft',
-      images: [],
-      schedule: [],
-      faqs: [],
-    },
-    validate: {
-      title: requiredField,
-      dateTime: requiredField,
-      type: requiredField,
-      location: locationValidation,
-      capacity: capacityValidation,
-    },
-  })
-
-  const onSubmit = async (errors, values) => {
+  const onSubmit = async (errors, values, isValid) => {
+    console.log(errors, values, isValid)
     const { location, dateTime, capacity, ...data } = values
+
     const eventData = {
       dateTime: dateTime ? dateTime.toISOString() : '',
       capacity: parseInt(capacity, 10),
@@ -50,7 +20,7 @@ const CreateEventScreen = () => {
       ...data,
     }
 
-    if (formState.isValid) {
+    if (isValid) {
       notifications.show({
         title: 'Complete los campos obligatorios',
         color: 'red',
@@ -82,7 +52,7 @@ const CreateEventScreen = () => {
     <>
       <Title>Creación de un evento</Title>
       <Space h={24} />
-      <EventForm formState={formState} onSubmit={onSubmit} />
+      <EventForm onSubmit={onSubmit} />
     </>
   )
 }
